Move strictyaml schema to validation module
