+++ openprescribing/data/measures/measures.py
@@ -1,27 +1,10 @@
 from django.conf import settings
 from strictyaml import (
-    Any,
-    Int,
-    Map,
-    MapCombined,
-    Optional,
-    Seq,
-    Str,
     YAMLValidationError,
     load,
 )
 
-
-class MeasureValidationError(Exception):
-    def __init__(self, measure_name, validation_error):
-        self.measure_name = measure_name
-        self.validation_error = validation_error
-        super().__init__()
-
-    def __str__(self):
-        return (
-            f"Measure '{self.measure_name}' failed to validate: {self.validation_error}"
-        )
+from .validation import MeasureValidationError, schema
 
 
 def load_measure(measure_name):
@@ -49,65 +32,3 @@ def all_measure_details():
             }
         )
     return measure_details
-
-
-# At the time of writing, this schema only validates measure features which have
-# been implemented.
-def schema():
-
-    NUMERATOR_KEY = "numerator"
-    DENOMINATOR_KEY = "denominator"
-
-    # Require the specified keys, but also accept any other keys
-    metadata = MapCombined(
-        {
-            "title": Str(),
-            "tags": Seq(Str()),
-            "why_it_matters": Str(),
-        },
-        Str(),
-        Any(),
-    )
-
-    # nb. schema does not validate this is a valid BNF code
-    bnf_code = Str()
-    bnf_code_list = Seq(bnf_code)
-    bnf_codes = Map(
-        {
-            Optional("included"): bnf_code_list,
-            Optional("excluded"): bnf_code_list,
-        }
-    )
-    # all of these query params are ANDed together
-    query_params = Map(
-        {
-            Optional("bnf_codes"): bnf_codes,
-            # corresponds to `cd` in `Ing`
-            Optional("ingredient_ids"): Seq(Int()),
-            # from a set list like `tablet`
-            Optional("forms"): Seq(Str()),
-            # can be `all`, `generic` or `branded`
-            Optional("product_type"): Str(),
-            # from a set list like `oral`
-            Optional("routes"): Seq(Str()),
-            # from a set list like `tablet.oral`
-            Optional("form_routes"): Seq(Str()),
-        }
-    )
-    # Currently this must be `items` or `list_size`
-    output = Str()
-    query = Map(
-        {
-            NUMERATOR_KEY: query_params,
-            Optional(DENOMINATOR_KEY): query_params,
-        }
-    )
-    schema = Map(
-        {
-            "metadata": metadata,
-            "output": Map({NUMERATOR_KEY: output, DENOMINATOR_KEY: output}),
-            "queries": Seq(query),
-        }
-    )
-
-    return schema
